refactor: 누락된 final 키워드 추가

+++ src/test/java/kitchenpos/support/MenuFixture.java
@@ -10,7 +10,7 @@ import kitchenpos.product.domain.Product;
 
 public class MenuFixture {
 
-    private static final int QUANTITY = 1;
+    private static final Long QUANTITY = 1L;
 
     public static MenuRequestDto 메뉴_생성(final String name,
                                        final BigDecimal price,
